-Additional work on shader compilation to support shader structures reflection "XSC only for now". -Rename "ReflectedShaderResource" to "ShaderResource". -Work on Material class to support LLGL. -Fixed some warnings.

+++ Nuclear Engine/Source/Engine/Assets/Mesh.cpp
@@ -76,7 +76,7 @@ namespace NuclearEngine {
 				desc.size = (unsigned int)data.indices.size();
 				mIB = Graphics::Context::GetRenderer()->CreateBuffer(desc, data.indices.data());
 			}
-			mIndicesCount = data.indices.size();			
+			mIndicesCount = static_cast<Uint32>(data.indices.size());			
 
 			data.Positions.clear();
 			data.UV.clear();
